pyproject.toml
[tool.poetry]
name = "botonera"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
python-dotenv = "^0.13.0"
pynacl = "^1.3.0"
ffmpeg = "^1.4"
aiohttp = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
